better organize code
