+++ cql2pgjson/src/main/java/org/folio/cql2pgjson/tbd/UnicodeIgnoreAccents.java
@@ -1,20 +1,20 @@
-package org.z3950.zing.cql.cql2pgjson;
+package org.folio.cql2pgjson.tbd;
 
 import java.util.Map;
 
 /**
- * Provides regular expressions for matching Unicode characters ignoring case.
+ * Provides regular expressions for matching Unicode characters ignoring accents/diacritics.
  */
-public final class UnicodeIgnoreCase {
-  private static Map<Character,String> unmodifiableMap = Unicode.readMappingFile("UnicodeIgnoreCase");
+public final class UnicodeIgnoreAccents {
+  private static Map<Character,String> unmodifiableMap = Unicode.readMappingFile("UnicodeIgnoreAccents");
 
-  private UnicodeIgnoreCase() {
+  private UnicodeIgnoreAccents() {
     throw new UnsupportedOperationException("Cannot instantiate utility class.");
   }
 
   /**
    * For each Character c the map provides a regexp that matches any
-   * String that is a case ignoring equivalent to c including c.
+   * String that is a accents/diacritics ignoring equivalent to c including c.
    * The map returns null if there is no other equivalent for c.
    * The map is unmodifiable.
    * @return  the map
