+++ cql2pgjson/src/main/java/org/folio/cql2pgjson/util/CqlUtils.java
@@ -2,9 +2,6 @@ package org.folio.cql2pgjson.util;
 
 /**
  * Help methods related to CQL syntax.
- *
- * @author hji
- *
  */
 public class CqlUtils {
