+++ cql2pgjson/src/test/java/org/z3950/zing/cql/cql2pgjson/Cql2SqlUtilTest.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+import org.folio.cql2pgjson.util.Cql2SqlUtil;
 import org.junit.Test;
 import org.junit.runner.RunWith;
