+++ cql2pgjson/src/test/java/org/z3950/zing/cql/cql2pgjson/SchemaTest.java
@@ -8,7 +8,8 @@ import static org.junit.Assert.*;
 import java.io.IOException;
 
 import org.folio.cql2pgjson.exception.QueryValidationException;
-import org.folio.cql2pgjson.exception.SchemaException;
+import org.folio.cql2pgjson.tbd.Schema;
+import org.folio.cql2pgjson.tbd.SchemaException;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
