+++ cql2pgjson/src/test/java/org/z3950/zing/cql/cql2pgjson/UnicodeIgnoreAccentsTest.java
@@ -4,6 +4,8 @@ import static org.junit.Assert.*;
 
 import java.util.Arrays;
 
+import org.folio.cql2pgjson.tbd.Unicode;
+import org.folio.cql2pgjson.tbd.UnicodeIgnoreAccents;
 import org.junit.Test;
 import org.junit.runner.RunWith;
