+++ cql2pgjson/src/test/java/org/z3950/zing/cql/cql2pgjson/UnicodeIgnoreNoneTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.*;
 
 import java.util.Arrays;
 
+import org.folio.cql2pgjson.tbd.Unicode;
 import org.junit.Test;
 import org.junit.runner.RunWith;
