+++ cql2pgjson/src/test/java/org/z3950/zing/cql/cql2pgjson/UnicodeMapFileGeneratorTest.java
@@ -6,6 +6,7 @@ import java.io.IOException;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
 
+import org.folio.cql2pgjson.tbd.UnicodeMapFileGenerator;
 import org.junit.Test;
 
 public class UnicodeMapFileGeneratorTest {
